refactor(auth): rename authLifecycle to lifecycle
