+++ src/app/providers/AppProviders.tsx
@@ -1,7 +1,7 @@
 import { useEffect, type PropsWithChildren } from "react";
 
 import { RemoteReadProvider } from "@/app/providers/remote-read";
-import { startAuthSession } from "./auth/authLifecycle";
+import { startAuthSession } from "./auth/lifecycle";
 
 export const AppProviders = ({ children }: PropsWithChildren) => {
   useEffect(startAuthSession, []);
